Create ProjectVacay
Welcome to my first project. This app allows individuals to plan their trip to defined destinations. The app provides discounts on car rentals based on the duration the individual will be away. When the individual inputs the required information, the app calculates the total one will need for the trip.